+++ tkviews/app.py
@@ -7,10 +7,12 @@ from pyviews.binding import use_binding
 from pyviews.code import run_code
 from pyviews.containers import get_container_pipeline, get_view_pipeline, get_for_pipeline, \
     get_if_pipeline
+from pyviews.presenter import get_presenter_pipeline
 from pyviews.rendering import RenderingPipeline, use_rendering, get_child_context
 from pyviews.rendering.pipeline import use_pipeline
 from pyviews.rendering.views import render_view
 
+from tkviews.canvas import get_canvas_pipeline
 from tkviews.core.rendering import TkRenderingContext, get_tk_child_context
 from tkviews.styles import get_style_pipeline, get_styles_view_pipeline
 from tkviews.widgets import get_root_pipeline, get_widget_pipeline, Root
@@ -32,6 +34,7 @@ def use_tkviews_pipelines():
     use_pipeline(get_root_pipeline(), 'tkviews.Root')
     use_pipeline(get_widget_pipeline(), 'tkinter')
     use_pipeline(get_widget_pipeline(), 'tkinter.ttk')
+    use_pipeline(get_presenter_pipeline(), 'tkviews.PresenterNode')
 
     use_pipeline(get_container_pipeline(), 'tkviews.Container')
     use_pipeline(get_view_pipeline(), 'tkviews.View')
@@ -41,6 +44,7 @@ def use_tkviews_pipelines():
     use_pipeline(get_style_pipeline(), 'tkviews.Style')
     use_pipeline(get_styles_view_pipeline(), 'tkviews.StylesView')
     use_pipeline(get_ttk_style_pipeline(), 'tkviews.TtkStyle')
+    use_pipeline(get_canvas_pipeline(), 'tkviews.canvas')
 
     use_pipeline(RenderingPipeline(pipes=[run_code]), 'tkviews.Code')